+++ python/tests/files_test.py
@@ -20,7 +20,6 @@ from acl_anthology.files import (
     AttachmentReference,
     PDFReference,
     VideoReference,
-    PapersWithCodeReference,
 )
 
 
@@ -61,58 +60,6 @@ test_cases_video = (
         "https://vimeo.com/385504611",
         "https://vimeo.com/385504611",
         False,
-    ),
-)
-
-
-test_cases_pwc = (
-    (
-        (
-            '<pwcdataset url="https://paperswithcode.com/dataset/wlasl">WLASL</pwcdataset>',
-        ),
-        None,
-        False,
-        [("WLASL", "https://paperswithcode.com/dataset/wlasl")],
-    ),
-    (
-        (
-            '<pwccode url="https://github.com/lksenel/coda21" additional="false">lksenel/coda21</pwccode>',
-        ),
-        ("lksenel/coda21", "https://github.com/lksenel/coda21"),
-        False,
-        [],
-    ),
-    (
-        (
-            '<pwcdataset url="https://paperswithcode.com/dataset/commonsenseqa">CommonsenseQA</pwcdataset>',
-            '<pwcdataset url="https://paperswithcode.com/dataset/qasc">QASC</pwcdataset>',
-            '<pwcdataset url="https://paperswithcode.com/dataset/squad">SQuAD</pwcdataset>',
-            '<pwcdataset url="https://paperswithcode.com/dataset/sciq">SciQ</pwcdataset>',
-        ),
-        None,
-        False,
-        [
-            ("CommonsenseQA", "https://paperswithcode.com/dataset/commonsenseqa"),
-            ("QASC", "https://paperswithcode.com/dataset/qasc"),
-            ("SQuAD", "https://paperswithcode.com/dataset/squad"),
-            ("SciQ", "https://paperswithcode.com/dataset/sciq"),
-        ],
-    ),
-    (
-        (
-            '<pwccode url="https://github.com/thunlp/OpenPrompt" additional="true">thunlp/OpenPrompt</pwccode>',
-            '<pwcdataset url="https://paperswithcode.com/dataset/glue">GLUE</pwcdataset>',
-        ),
-        ("thunlp/OpenPrompt", "https://github.com/thunlp/OpenPrompt"),
-        True,
-        [("GLUE", "https://paperswithcode.com/dataset/glue")],
-    ),
-    (
-        # This happens, so it needs to be handled
-        ('<pwccode url="" additional="true"/>',),
-        None,
-        True,
-        [],
     ),
 )
 
@@ -171,30 +118,6 @@ def test_video_reference_to_xml(xml, name, url, permission):
     ref = VideoReference(name=name, permission=permission)
     assert ref.url == url
     assert etree.tostring(ref.to_xml("video"), encoding="unicode") == xml
-
-
-@pytest.mark.parametrize("xml_list, code, community_code, datasets", test_cases_pwc)
-def test_pwc_reference_from_xml(xml_list, code, community_code, datasets):
-    ref = PapersWithCodeReference()
-    for xml in xml_list:
-        element = etree.fromstring(xml)
-        ref.append_from_xml(element)
-    assert ref.code == code
-    assert ref.community_code == community_code
-    assert ref.datasets == datasets
-
-
-@pytest.mark.parametrize("xml_list, code, community_code, datasets", test_cases_pwc)
-def test_pwc_reference_to_xml(xml_list, code, community_code, datasets):
-    ref = PapersWithCodeReference(
-        code=code,
-        community_code=community_code,
-        datasets=datasets,
-    )
-    actual_xml_list = ref.to_xml_list()
-    assert len(xml_list) == len(actual_xml_list)
-    for expected_xml, actual_xml in zip(xml_list, actual_xml_list):
-        assert etree.tostring(actual_xml, encoding="unicode") == expected_xml
 
 
 def test_reference_cant_change_template_field():